requirements.txt
djangorestframework-simplejwt
Django==4.0.8
PyYAML==6.0
django-filter==22.1
pydantic==1.9.2
djangorestframework==3.13.1
celery==5.2.7
psycopg2-binary==2.9.3
python-dotenv==0.20.0
drf-yasg==1.20.0
djoser==2.1.0

pytest==7.1.2
pytest-django==4.5.2
pytest-cov==4.0.0
pytest-xdist==2.5.0
pytest-factoryboy
factory-boy==3.2.1
freezegun==1.2.2